tweaks and clippy suggestions

+++ vm/src/lib.rs
@@ -52,7 +52,7 @@ impl VM {
 
             Opcode::LOAD => {
                 let dest = self.next_byte() as usize;
-                let val = self.next_value() as i32; // i16 gets interpreted as i32
+                let val = self.next_value();
 
                 self.registers[dest] = val;
             }
@@ -127,8 +127,8 @@ impl VM {
         byte
     }
 
-    fn next_value(&mut self) -> i16 {
-        ((((self.next_byte() as u16) << 8) | self.next_byte() as u16) as u16) as i16
+    fn next_value(&mut self) -> i32 {
+        (((self.next_byte() as u16) << 8) | self.next_byte() as u16) as i32
     }
 
     fn decode_opcode(&mut self) -> Opcode {
@@ -163,13 +163,12 @@ mod tests {
     fn grab_2_bytes() {
         let mut vm = VM::with_program(vec![20, 66, 4, 8]);
 
-        assert_eq!(vm.next_value(), 5186i16);
+        assert_eq!(vm.next_value(), 5186i32);
         assert_eq!(vm.pc, 2);
 
-        assert_eq!(vm.next_value(), 1032i16);
+        assert_eq!(vm.next_value(), 1032i32);
         assert_eq!(vm.pc, 4);
     }
-
     #[test]
     fn opcode_load() {
         let mut vm = VM::with_program(vec![
